use late import for ci logging function
